Revert "Code cleanup of routes config loader and API server (#424)"
This reverts commit 1ee3eb4de381bb2b03c3468f167a7c681ea7d54c.
