+++ .github/workflows/release.yml
@@ -10,7 +10,7 @@ jobs:
   release:
     uses: itzg/github-workflows/.github/workflows/go-with-releaser-image.yml@main
     with:
-      go-version-file: 'go.mod'
+      go-version: "1.24.4"
       enable-ghcr: true
     secrets:
       image-registry-username: ${{ secrets.DOCKERHUB_USERNAME }}
