+++ go.mod
@@ -1,6 +1,8 @@
 module github.com/itzg/mc-router
 
-go 1.24.4
+go 1.24.0
+
+toolchain go1.24.4
 
 require (
 	github.com/fsnotify/fsnotify v1.9.0
